Reland "updatable-via-apex on c2 manifest"
c2 is in an APEX, but it's not declared in the manifest.
This could block us from updating c2 on old devices that
receive this mainline module update, so putting a time
horizon on that issue.

This reverts commit 206960e4bc73d80947e5deaabed12cdd94090459.

Reason for revert: re-place, should use APEX name.

This was reverted because I accidentally used the old style
'updatable-in-apex="true"' rather than specifying the APEX
name. The test that caught this issue is added into
TEST_MAPPING and verified locally, so that we are sure it
doesn't cause a failure again.

We still need to merge this as without this change, updating
the version of c2 in the APEX will break updating devices.
So, this prevents that issue from happening in the future.

Change-Id: I9211514eb475d254347977d4eae2338841e86e9d
Test: vts_treble_vintf_framework_test (needed)
Bug: 385355208