audio policy: optimize Bluetooth device switch
Add an argument to AudioPolicyManager::setDeviceConnectionState()
to indicate that a device disconnection is part of a disconnect/connect
sequence corresponding to a switch betwen two devices of the same type
(same profile for Bluetooth devices).
This is currently only used for Bluetooth devices.
This allows to skip some of the rerouting operations that normally
happen when a device is disconnected knowing that the route evaluation
will happen again when the new device is connected.

Flag: com.android.media.audio.optimize_bt_device_switch
Bug: 373867402
Test: Bluetooth audio use cases regression
Change-Id: I6db94be4cd16804477ca6b1cbaf46015c06ca8c7