Remove bad logic to manage when working on next sdk
Code added to cover the interval where we're working on the next
release but have not bumped the SDK is bad; it was always bumping.
Drop that adjustment code so it only considers the current sdk, no
attempt to figure out if this build is during 'developing next
release'.

Bug: 388925029
Bug: 242057615
Test: 'dumpsys media.player', check APV codec status on U/V/B
Flag: EXEMPT bugfix
Change-Id: Ieb8235347c68ccd0826888f315e0089bf8e35b3c