Add move only type erased function
std::function does not permit move only function objects, which is
necessary for convenient usage of std::packaged_task. move_only_function
is being added in a subsequent version of the standard.

Add a Runnable class which is a limited move only no-param function
object with SBO only for typical double pointer types, which also erases
the return value of the type. This is useful for constructing executors.

Test: atest runnable_tests
Flag: EXEMPT util
Bug: 391668615
Change-Id: I44c25992465956e33944060cc0dcc07a768cb768