Wrap statsd calls with SDK checks
protect mediametrics calls to statsd API with SDK thresholds.
Prevent invocation of an API on a release where it doesn't exist.
This is mostly for tooling to ensure safety; the
module-based mediametrics won't be enabled on devices earlier
than when statsd is available. If that was somehow changed,
this change would prevent jumping to the not-present routines;
the side effect would be no metrics logged in such a situation.

Bug: 189976186
Flag: EXEMPT refactor
Test: atest CtsMediaHostTestCases
Change-Id: I2efda592e927d5c73c981378bb94db15645d722d