[audio] Avoid startOp when restrictions in place
When AppOp restrictions are in place, starting an op will return
SOFT_DENY, but leave the op in a paused state, which can restart
unexpectedly when the restriction is removed (despite no data delivery
occurring).

Work around this by first checking if recording is restricted, and
avoiding the start call in that case.

Test: manual: start and stop recording with toggle restriction
Test: atest CtsMediaAudioPermissionTestCases
Fixes: 399138369
Bug: 293603271
Flag: EXEMPT bugfix
Change-Id: Ie7d2661b4027b5ce6bee9b6c63b70f0b120dcbcc