audioflinger: add support for GMAP mode
Add support for insertion of special tag in playback
track metadata to indicate request for bidirectional mode to the
BT stack for GMAP mode.
If an apk has been recording audio in the past and plays for game
usage, bidirectional mode is assumed.
This is a short term work around until a proper API is introduced
for explicit indication of bidirectional request by the app.

Bug: 366456949
Test: make
Flag: com.android.media.audioserver.enable_gmap_mode
Change-Id: Ib490660dc9af6c808ba91b154ae4714a83f28bc6