Reject start other output if bit-perfect output is active.
When bit-perfect output is active, all clients should be routed to
bit-perfect output if the selected output devices contain the
bit-perfect output device. In that case, when trying to start output on
other playback thread with the bit-perfect device, the start request
should fail to allow the client to reattach to the right playback
thread.

Bug: 391116315
Test: atest audiopolicy_tests
Flag: EXEMPT bugfix
Change-Id: Iac6a8621c9680908dce3628fe216086113605e2f