stagefright: distinguish HAL name from name in MediaCodecInfo
We no longer allow multiple supported media types
for a media codec info, but this change was done at the
NDK level, which created differing names in SDK and NDK.
Make this change at the MediaCodecList level, so
SDK and NDK uses the same codec name.

For this we need to track the codec name separately
from the HAL name.

Flag: EXEMPT bugfix
Test: manual test with ag/32415283
Test: atest android.mediav2.cts.NativeAMediaCodecInfoTest 
Bug: 401935274
Change-Id: I15c73fa4bfbd8077b50ea934ddc918202f6c9f69