Use mix port id when opening stream.
Starting from AIDL HAL, the HAL reports device capabilities via the HAL
API. This makes it possible to have a synchronized port id between the
framework and the HAL. In that case, it makes sense to use port id
provided by the HAL when opening stream. That can improve the logic when
looking for the right mix port to open stream.

Bug: 253474154
Bug: 354751111
Test: atest audiopolicy_tests
Test: atest CoreAudioHalAidlTest
Test: atest AudioTrackTest AudioRecordTest on Pixel 7 and 9
Flag: EXEMPT bugfix
Change-Id: I9efb9ad465fcd6c72e0e0e13f3db5624b16b545e