Fix for the crash in Audioflinger
Error:SIGABRT:libaudioflinger.so:variant:android::AudioFlinger::TrackHandle::getCblk(std::__1::optional<android::media::SharedFileRegion>*):libc.so:l

Summary: We observed this crash on quest devices but hard to reproduce.
Reviewed available data from reports and investigated the code. The
problem appears to be that the C++ code does not check for the proper
data type at run time, instead directly dereferencing the callback type.
This is acceptable for normal cases, but in failure scenarios, it can
result in a crash or abort exception.

Flag: EXEMPT bugfix
Test: atest AudioTrackTest
Bug: 408222100
Change-Id: Ic35e27cc44424f9093dfa8c04b71aac2da879363
Signed-off-by: Abhishek Gadewar <abhishekgadewar@fb.com>