c2SoftAacEnc: fix output timestamps for partial inputs
In scenarios where enqueued input is not an integral multiple of aac
packet size, the encoder consumes the input and returns output for all
samples except for the (input modulus aac_packet_size). These samples
are buffered inside the library and waited till the next enqueue call is
made or eos is signalled. In these cases, the output timestamp should
consider the already batched samples in its computations. This is not
managed correctly and this is now fixed.

Bug: 401377715
Test: atest MctsMediaV2TestCases:AudioEncoderTest
Test: atest MctsMediaV2TestCases:CodecEncoderTest
Test: atest MctsMediaEncoderTestCases:EncoderTest
Test: atest MctsMediaV2TestCases:CodecEncoderBlockModelTest
Test: atest MctsMediaV2TestCases:CodecEncoderMultiAccessUnitTest
Test: atest MctsMediaV2TestCases:CodecEncoderBlockModelMultiAccessUnitTest
Test: atest VtsHalMediaC2V1_0TargetAudioEncTest
Flag: EXEMPT bugfix

Change-Id: Iac418cacb1064a2e900f7e11909db28e93fb7f66