AudioFlinger: Support Track container polymorphism
The ThreadBase class needs to directly access TrackBase objects
to support better reuse of code, and eventually Mmap internal
support.

This CL begins the migration of mActiveTracks and mTracks
to hold TrackBase objects, and then the migration to ThreadBase.

We create views on mActiveTracks to the actual IAfTrack, IAFRecordTrack
interfaces to simplify and make reliable the calling
from PlaybackThread and RecordThread as needed.  We note that
IAfTrackBase may be inherited through virtual base classes -
casting to base is okay, but to derived cannot be determined
in static fashion.

Flag: EXEMPT bugfix
Test: atest CtsMediaAudioTestCases
Test: atest CtsNativeMediaAAudioTestCases
Bug: 410038399
Change-Id: Ie6fdf61293bfaf5ea6d3897344d5a72f01b07666