Do not fail when an input surface is empty.
Before when an empty view Surface was passed getUniqueID failed and the
function did not execute. Passing an empty surface is valid so we have
to allow for this behavior. The current version uses 0 for the ID for an
empty surface and fills it in correctly if it's not empty.

Bug: 412428300
Test: atest android.security.cts.CameraPermissionTest#testAppStreaming1_softDenial_cameraMute
Flag: com.android.graphics.libgui.flags.wb_libcameraservice
Change-Id: I419eea1b80cda63046861a9e8744b67941c9f792