screenrecord: Clean up the encoder before the overlay.
There's a chance the overlay might be caught waiting for a dequeuable
buffer from the encoder and the overlay thread might be locked up. The
surface the overlay is waiting for is the encoder's input surface.

If we clean up the encoder first, it should disconnect that surface and
fail the blit, which will keep everything from being locked up.

Bug: 386318952
Flag: EXEMPT test/command line tool, small change
Test: Ran some ScreenRecorderRule tests, manually ran w/ overlay

Change-Id: I6a045cecf7194dc46e4f89b88fc44431173013be