Reapply "APM: use routing checks with ioprofile"
This reverts commit dc1aa1bb177c70ada8a2a08ad35ebd4624e00b38,
and adds the change to enhance flag checking.

In the original CL, when the flag is disabled,
|getAudioMixPort| will return |BAD_VALUE| if the IO handle is none.

That makes some routable checks end up in false, and break CUJs.

This CL adds the flag check when such checks are involved, so that
the legacy logic is truly guarded.

Bug: 367117623
Test: Verify BT audio works on incoming call w&w/o flag on Caiman
Flag: com.android.media.audioserver.enable_strict_port_routing_checks

Change-Id: If3b439f842e0c47bcf19fceab8f5d7829271576e