Add getEffectFromUuid_l to EffectChain
Add a function called getEffectFromUuid_l to ask if a specific effect is
available in an effect chain. There is already a similar function that
does the same but for a type, and actually only get the first effect
with that type even if there are several. It is more exact to ask for a
specific UUID.

Flag: EXEMPT bugfix
Test: m
Bug: 413628250
Change-Id: Icbe3a46d6b43343465f61b191923e67652a012b0