audio:Fix the accumulation bug of mFramesWritten
Assume that the currently requested frameCount is 1920, but the current sharebuffer has only 1800 available space. The first loop will copy it to the shared buffer. The second loop will directly return the obtained buffer due to the lack of available space, and will not add 1800.

mFramesWritten will assign a timestamp variable mFramesWrittenServerOffset in restoreTrack_l, and get it in getTimestamp_l.

Flag: EXEMPT bugfix
Bug: 414543147
Test: See bug
Change-Id: Ie80fdf7000a850f600bc178fc1efc7260a2b0aa4
Signed-off-by: jiangyao <jiangyao@xiaomi.com>