C2SoftVideoDec: Handle process calls with csd and picture
CodecSpecificData and access units corresponding to pictures are
expected to be sent separately. If they are sent together it is seen
that software components fail to decode.

For inputs with csd and picture, the decoder first processes the csd.
This may raise a config update request. This request is generally
communicated to the C2 framework via the work item through which the csd
is received. But as there is more data to process (data belonging to
picture), the decoder continues to decode. This may result in an output
frame. Due to reorder depth restrictions, if this frame is not sent by
the decoder, then the work item will not be returned, thereby not
returning the config update request as well. If this config update
request is to update output delay information, then C2 will send more
inputs. As this update never happened, the C2 is expecting an output and
component is expecting more inputs. This change addresses this lock.

This change is restricted to components that require csd

Bug: 320448605
Test: atest MctsMediaV2TestCases
Test: atest MctsMediaDecoderTestCases
Test: atest VtsHalMediaC2V1_0TargetVideoDecTest
Flag: EXEMPT bugfix
Change-Id: I8145bbaf453c1021dd78f589ce3bde48cfd2b0cc