[audio] Fix missing package Attrs from mediaserver
To correctly tolerate missing packages in attribution sources from
trusted UIDs (typically mediaserver), when those UIDs are playing audio
on behalf of a client, we should fill the package name based on the UID
in the attribution, not the UID of the caller.

This fixes AppOps failures due to package name mismatch behind the new
permission logic flag, due to incorrectly filling in the package as
"media" with a non mediaserver UID, which result in incorrectly silenced
audio tracks.

Bug: 411348062
Flag: EXEMPT bugfix
Test: manual repro with app using opensles non-pcm
Change-Id: I11eb2054aef135779ddf8f1e18e3bd553321e1e5