Allocate buffer from heap to copy data when exiting standby.
When the stream is an offload stream, there may be lots of data written
but not played when the stream is in standby mode. When exiting standby,
aaudio will have to copy those data to the new shared buffer. Allocate
the buffer from heap to copy data to avoid crash as the data size may be
greater than the stack limitation.

Bug: 416107227
Test: restart offload stream from standby
Flag: EXEMPT bugfix
Change-Id: I12140819a360044014d604ca898c66c710ac9729