DynamicsProcessing: Update AIDL Conversion Logic
Modify the AIDL conversion to read the second boolean value from the input configuration, representing the mEnabled parameter, instead of the previously used mInUse value. Additionally, remove the check that restricted enabling Equalizer and MBC bands when the channel configuration was not explicitly set to true.

This change aligns the AIDL behavior with the existing HIDL implementation.

Flag: EXEMPT bugfix
Bug: 397675206
Test: atest VtsHalDynamicsProcessingTargetTest
Change-Id: I83a97767542a3b1202df51c0c2f1d3636063abd9