cameraserver: Get surface format from BufferQueue instead of OutputConfiguration
The cached value of OutputConfiguration's cached format may not match the Surface's
configured format as a result of format overrides in SurfaceUtils.getSurfaceFormat() -
b/9487482 for historical details.
Changing SurfaceUtils.getSurfaceFormat() may cause a slew of assumptions made by framework
and camera extensions code. So we revert back to the old behavior of directly querying the BufferQueue for the format instead of relying on OutputConfiguration.

b/417531963 will follow up on whether the avoiding the override is feasible.

Bug: 416270160

Test: Camera CTS, Virtual devices camera CTS

Flag: EXEMPT; bug fix
Change-Id: I70a22a28b52ee860c704fdffe4b3308108ff679b