Allow decoders to be configured with channel mask
For decoders like IAMF where the output layout could be configured by the user of the decoder, allow the channel mask to be set in configure and also reported back by the decoder.

Users of decoders can currently safely set both KEY_MAX_OUTPUT_CHANNEL_COUNT (SDK >= 32) and KEY_CHANNEL_MASK to be able to compatible with the codec framework without and with this change.

Bug: 388336285
Flag: android.media.swcodec.flags.iamf_software_decoder
Change-Id: I4f182bd2a6b236253b88c46320c1aff92f92ca84