Release component on destructor if needed
Currently Component::release() is only called in Component::OnBinderDied(). In some cases, OnBinderDied() is not triggered, resulting in component not released and potential memory leak.

To avoid this issue, keep track of the release status for a component, and call reset() and release() on destructor if it has not been released.

Bug: 405001050
Test: atest DecoderTest, verified by AMLogic
Flag: EXEMPT bugfix
Change-Id: I601b01a8e3627e552927d37d23378c29e271c3d1