Suspend callback thread if the buffer is full for offload playback.
For offload playback, the shared mmap buffer is large so that there is
no needed to wake up often to get data from client if the buffer is
almost full. We can suspend the callback thread to drain the data and
wakeup when it is closed to empty buffer.

Bug: 409798445
Test: OboeTester, start offload playback
Test: atest AAudioTests
Test: test_pcm_offload -f2 -c2 -r48000
Flag: com.android.media.audioserver.mmap_pcm_offload_support
Change-Id: Ib17f914932169edba0d2f864377ca1bda4488a7e