Implement forced removal of direct outputs on HAL stream hard error
Whenever an asynchronous output reports a "hard" (unrecoverable)
error, close the corresponding APM output and AF I/O thread
and release its clients as a single operation by the APM.

Removing clients track-by-track leaves a chance for the client
to restart playback on the output which is being closed.
By making closing a single APM operation this possibility is
eliminated.

This is currently only implemented for direct streams to avoid
dealing with complications of reopening of non-direct outputs.

Flag: EXEMPT bugfix
Bug: 414259900
Test: atest audiopolicy_tests
      manual with hard error simulation from the HAL
Change-Id: I459b0b0adc7f14011da09f1889ec9e45491f1b8e