APM: improve getAudioMixPort simulation in test
The unit tests failed to catch the regression when flag is disabled due
to that it has mocked the critical function that contains the flag.

This change makes the mocked function respect the flag and produce
similar results so that it will result in failure without the fix.

Bug: 417874384
Test: Verify test fails when flag disabled
Flag: com.android.media.audioserver.enable_strict_port_routing_checks

Change-Id: I180618b4cc14c164c9db7d163e2890b86d06a26b