Add fsetxattr to allowed syscall list
Starting with API level 202604, libcutils will be using memfds for
native code. When a memfd is created via libcutils, the memfd is
assigned a different security context. This is a temporary workaround
until selinux can be augmented to distinguish between memfds and tmpfs
files.

Since fsetxattr() is not part of the seccomp policies for the
mediacodec and mediaswcodec processes, the invocation to that function
in libcutils causes those processes to crash. Therefore, add
fsetxattr() to those secccomp policies to fix this and use memfd.

Flag: EXEMPT bugfix
Bug: 426634646
Change-Id: I8809f75f09990d594da4b0918e5dbc68e43f6ef7
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>