C2SoftOpusEnc: fix pts offset for first output frame
While computing output timestamp of current frame, number of processed
samples is incremented and then used. This is incorrect and would result
in next frame's timestamp to be marked as current. This is addressed.
Also, at eos, it is seen that a mute buffer is encoded and sent always.
This is not required. Only when the encoder plugin context has left over
unprocessed samples, these are padded to the frame duration using
silence and encoded. This too is handled.

Bug: 139484367
Test: atest MctsMediaV2TestCases -- --module-arg \
MctsMediaV2TestCases:instrumentation-arg:media-type-sel:=opus
Test: atest MctsMediaEncoderTestCases:EncoderTest -- --module-arg \
MctsMediaEncoderTestCases:instrumentation-arg:media-type-prefix:=audio/opus
Test: atest VtsHalMediaC2V1_0TargetAudioEncTest
Flag: EXEMPT bugfix

Change-Id: I85f1a6c8673f271954fe51a600167993cc0f119c