Add AIDL CameraCompatMode.
Existing int (representing enum) rotationOverride contains multiple pieces of information: by how much to rotate in rotate-and-crop, and whether to change sensor orientation, making it difficult to parse and extend.
Refactoring to use CameraCompatMode data object (aidl) allows sending more information, in a more readable format. It also ensures that no more refactorings of this variable is needed if more data needs to be sent (originally, rotationOverride was a boolean, and transition required the same amount of refactoring).

AIDL is needed to pass data cross-languages - protos are not supported in AIDL.

Flag: EXEMPT refactoring, which would be difficult to guard. No behaviour change.
Test: Manual, tested with and without camera compat mode.
Bug: 432651608
Bug: 365725400

Change-Id: I9ced90161ce179e3f2279a9a2d23327e371c8b1d