Move AppOps check off binder thread
We currently check the the AppOp state in response to a callback on the
binder thread which the callback is dispatched on.

However, this incoming callback is oneway, which means we don't get the
re-entrancy behavior from sync binder callback calls. This means we can
have a implicit deadlock against both of the binder threadpools:
all of our binder threads are attempting a binder txn on system server,
and all system server threads are attempting a binder txn on us (under
synthetic/extreme contention).

Fixes: 436769260
Bug: 380150343
Flag: EXEMPT bugfix
Test: manual background playback 
Change-Id: I6a6a6964dca618b20972a21faeacaac0465c8035