Use CameraCompatibilityInfo instead of rotationOverride.
Int (representing enum) rotationOverride contained multiple pieces of information: by how much to rotate in rotate-and-crop, and whether to change sensor orientation, making it difficult to parse and extend.
This refactoring allows sending more information, in a more readable format. It also ensures that no more refactorings of this variable is needed if more data needs to be sent (originally, rotationOverride was a boolean, and transition required the same amount of refactoring).

This change has started as CameraCompatMode which was planned to be copied from CameraCompatibilityInfo, but since using one implementation is possible, it will be more future-proof.

Flag: EXEMPT refactoring, which would be difficult to guard. No behaviour change.
Test: atest PropertyTest
Test: Manual, tested with and without camera compat mode.
API-Coverage-Bug: 436463722
Bug: 432651608
Bug: 365725400
Bug: 414347702

Change-Id: I06324ee70c90f9913545f8c05d3bef5daf07298d