Do not check routable when getting new devices for output.
The routability is checked when setting the output devices. When getting new
devices for output, it does not need to consider if the device is routable
or not. The function is used to pick up the possible routed devices based
on the phone mode and active use cases.

Bug: 437780606
Test: atest audiopolicy_tests
Test: Phone call via hearing aid device
Flag: EXEMPT bugfix
Change-Id: I7206c708240e8cd6e5ea9d50d0d2e6306c4ac9bb