Add empty pkg check for service AppOp exemption
As part of I5186e190bf52480f35436cce75abc8c506150fd9, audioserver no
longer checks if there are no package names under a uid before
exempting a non-app attribution with that uid for OP_PLAY_AUDIO checks
(which are used for Zen).

Although telecom is on the exemption list for Zen, it is not for effect
suppression (when zen is not enabled), so the implementation relied on
the silencing behavior from audioserver in that case.

Restore this functionality with the new package cache (which prevents
deadlock issues) to address the regression, at least until the
zen/telecom API is modified.

Bug: 413631827
Flag: EXEMPT bugfix
Test: manual via logs
Change-Id: I6a6a696497a957888346f301e34393483f320e31